+++ include/render/PathTracingIntegrator.h
@@ -116,6 +116,9 @@ namespace render {
     Colord sampleDirectLighting(const Scene& scene, const LightSampler& lightSampler,
                                 const HitPoint& hitPoint, const Material& material,
                                 const Vector3d& wi, State& state, int bounce) const;
+    Colord emittedRadiance(const LightSampler& lightSampler, const Material& material,
+                           const Rayd& ray, const HitPoint& hitPoint, bool sampledFromBsdf,
+                           double bsdfSamplePdf, bool bsdfSampleDelta) const;
     Colord directLighting(const Scene& scene, const Light& light, const HitPoint& hitPoint,
                           const Material& material, const Vector3d& wi, const Vector2d& lightSample,
                           State& state) const;